# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Primavera V10 API - LLM Ready",
    "version": "4.0.0",
    "description": "API REST otimizada para consumo por LLMs. Dados do ERP Cegid Primavera V10 da Cimenteira do Louro. Todos os cálculos de faturação usam 34 tipos de documento com notas de crédito como valores negativos.",
    "contact": {
      "name": "Cimenteira do Louro",
      "url": "http://primavera-api.cimenteiradolouro.local"
    }
  },
  "servers": [
    {
      "url": "http://primavera-api.cimenteiradolouro.local",
      "description": "Servidor Primavera API"
    }
  ],
  "security": [
    { "basicAuth": [] },
    { "apiKeyHeader": [] },
    { "apiKeyQuery": [] }
  ],
...
  "paths": {
    "/api/clientes": {
      "get": {
        "tags": ["Clientes"],
        "summary": "Listar clientes",
        "description": "Retorna lista paginada de clientes com pesquisa e filtros por zona e vendedor.",
        "operationId": "listarClientes",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Pesquisa por nome, código ou contribuinte do cliente",
            "schema": { "type": "string" }
          },
          {
            "name": "zona",
            "in": "query",
            "description": "Filtrar por zona geográfica",
            "schema": { "type": "string" }
          },
          {
            "name": "vendedor",
            "in": "query",
            "description": "Filtrar por código do vendedor",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número da página (começa em 1)",
            "schema": { "type": "integer", "default": 1, "minimum": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número de resultados por página",
            "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 500 }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ClienteResumo" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/artigos": {
      "get": {
        "tags": ["Artigos"],
        "summary": "Listar artigos",
        "description": "Retorna lista paginada de artigos com pesquisa e filtros por família e subfamília.",
        "operationId": "listarArtigos",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Pesquisa por código ou descrição do artigo",
            "schema": { "type": "string" }
          },
          {
            "name": "familia",
            "in": "query",
            "description": "Filtrar por família de artigos",
            "schema": { "type": "string" }
          },
          {
            "name": "subfamilia",
            "in": "query",
            "description": "Filtrar por subfamília de artigos",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número da página (começa em 1)",
            "schema": { "type": "integer", "default": 1, "minimum": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número de resultados por página",
            "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 500 }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de artigos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ArtigoResumo" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
    "/api/query": {
      "post": {
        "tags": ["Utilitários"],
        "summary": "Executar query SQL",
        "description": "Executa uma query SQL direta na base de dados do ERP Primavera. Usar com cuidado - apenas queries SELECT são permitidas.",
        "operationId": "executarQuery",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sql"],
                "properties": {
                  "sql": {
                    "type": "string",
                    "description": "Query SQL a executar (apenas SELECT)",
                    "example": "SELECT TOP 10 Cliente, Nome, Zona FROM Clientes WHERE Zona = @zona"
                  },
                  "params": {
                    "type": "object",
                    "description": "Parâmetros da query (previne SQL injection)",
                    "additionalProperties": true,
                    "example": { "zona": "Norte" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado da query",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "rowCount": { "type": "integer", "example": 10 },
                    "executionTime": { "type": "string", "example": "45ms" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Query inválida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": false },
                    "error": { "type": "string", "example": "Apenas queries SELECT são permitidas" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Autenticação HTTP Basic"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Chave de API via header"
      },
      "apiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apikey",
        "description": "Chave de API via query parameter"
      }
    },
    "schemas": {
      "ClienteResumo": {
        "type": "object",
        "properties": {
          "cliente": { "type": "string", "example": "C0001" },
          "nome": { "type": "string", "example": "Construções Silva Lda" },
          "contribuinte": { "type": "string", "example": "501234567" },
          "zona": { "type": "string", "example": "Norte" },
          "localidade": { "type": "string", "example": "Braga" },
          "vendedor": { "type": "string", "example": "V001" },
          "telefone": { "type": "string", "example": "+351 253 123 456" },
          "email": { "type": "string", "example": "geral@construcoessilva.pt" },
          "ativo": { "type": "boolean", "example": true }
        }
      },
...
      "ArtigoResumo": {
        "type": "object",
        "properties": {
          "artigo": { "type": "string", "example": "CIM001" },
          "descricao": { "type": "string", "example": "Cimento Portland CEM II 42.5R" },
          "familia": { "type": "string", "example": "Cimentos" },
          "subfamilia": { "type": "string", "example": "Portland" },
          "unidade": { "type": "string", "example": "TON" },
          "precoVenda": { "type": "number", "format": "double", "example": 12.00 },
          "iva": { "type": "number", "format": "double", "example": 23.0 },
          "ativo": { "type": "boolean", "example": true }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 50 },
          "total": { "type": "integer", "example": 1250 },
          "totalPages": { "type": "integer", "example": 25 },
          "hasNext": { "type": "boolean", "example": true },
          "hasPrev": { "type": "boolean", "example": false }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Não autorizado - credenciais inválidas ou em falta",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Autenticação necessária" },
                "code": { "type": "string", "example": "UNAUTHORIZED" }
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Erro interno do servidor",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": { "type": "boolean", "example": false },
                "error": { "type": "string", "example": "Erro interno do servidor" },
                "code": { "type": "string", "example": "INTERNAL_ERROR" }
              }
            }
          }
        }
      }
    }
  }
}